==> notebooks/01_process_challenge_set.ipynb <==
import pandas as pd

challenge_set_df = pd.read_csv('../PRC_data/challenge_set.csv')

display(challenge_set_df)

<h3>Drop the columns <strong>callsign, name_adep, and name_ades<strong/></h3>

challenge_set_df = challenge_set_df.drop(columns=['callsign', 'name_adep', 'name_ades'])

<h3>Split the <strong>date</strong> column into</h3>
<ul>
    <li>month_day</li>
    <li>month</li>
    <li>day</li>
</ul>
<h3>Drop the <strong>date</strong> columns</h3>

challenge_set_df['date'] = pd.to_datetime(challenge_set_df['date'])
challenge_set_df['month_day'] = challenge_set_df['date'].dt.strftime('%m-%d')  # Format to get month and day together
challenge_set_df['month'] = challenge_set_df['date'].dt.month  # Extract the month
challenge_set_df['day'] = challenge_set_df['date'].dt.day  # Extract the day

display(challenge_set_df[['flight_id', 'date', 'month_day', 'month', 'day']])

challenge_set_df = challenge_set_df.drop(columns=['date'])

<h3>Split the <strong>actual_offblock_time</strong> and <strong>arrival_time</strong> columns into</h3>
<ul>
    <li><strong>actual_offblock_time_hour_minute<strong/></li>
    <li><strong>actual_offblock_time_hour<strong/></li>
    <li><strong>actual_offblock_time_minute<strong/></li>
    <li><strong>arrival_time_hour_minute<strong/></li>
    <li><strong>arrival_time_hour<strong/></li>
    <li><strong>arrival_time_minute<strong/></li>
</ul>
<h3>Drop the <strong>actual_offblock_time</strong> and <strong>arrival_time</strong> columns</h3>

challenge_set_df['actual_offblock_time'] = pd.to_datetime(challenge_set_df['actual_offblock_time'])
challenge_set_df['actual_offblock_time_hour_minute'] = challenge_set_df['actual_offblock_time'].dt.strftime('%H:%M')
challenge_set_df['actual_offblock_time_hour'] = challenge_set_df['actual_offblock_time'].dt.hour
challenge_set_df['actual_offblock_time_minute'] = challenge_set_df['actual_offblock_time'].dt.minute

challenge_set_df['arrival_time'] = pd.to_datetime(challenge_set_df['arrival_time'])
challenge_set_df['arrival_time_hour_minute'] = challenge_set_df['arrival_time'].dt.strftime('%H:%M')
challenge_set_df['arrival_time_hour'] = challenge_set_df['arrival_time'].dt.hour
challenge_set_df['arrival_time_minute'] = challenge_set_df['arrival_time'].dt.minute

display(challenge_set_df[['flight_id', 
                          'actual_offblock_time', 'actual_offblock_time_hour_minute', 'actual_offblock_time_hour', 'actual_offblock_time_minute',
                          'arrival_time', 'arrival_time_hour_minute', 'arrival_time_hour', 'arrival_time_minute']])

challenge_set_df = challenge_set_df.drop(columns=['actual_offblock_time', 'arrival_time'])

<h3>Construct/Order challenge set</h3>

challenge_set_df = challenge_set_df[[
    'flight_id', 'month_day', 'month', 'day', 'actual_offblock_time_hour_minute', 'actual_offblock_time_hour', 'actual_offblock_time_minute',
    'adep', 'country_code_adep', 'arrival_time_hour_minute', 'arrival_time_hour', 'arrival_time_minute', 'ades', 'country_code_ades',
    'aircraft_type', 'wtc', 'airline', 'flight_duration', 'taxiout_time', 'flown_distance', 'tow'
]]

display(challenge_set_df)

<h3>The dataset is them checked for any Nan values in any cells</h3>

rows_with_nan = challenge_set_df[challenge_set_df.isna().any(axis=1)]
display(rows_with_nan)

<h3>Store the processed data</h3>

challenge_set_df.to_csv("./data/processed_challenge_set.csv", index=False)
print(f"./data/processed_challenge_set.csv is saved!")
