==> notebooks/4_add_trajectories_to_datasets.ipynb <==
import pandas as pd

<h3>Read summarized_trajectories</h3>

summarized_trajectory_df = pd.read_csv('summarized_trajectory.csv')
display(summarized_trajectory_df)

<h3>Add trajectories to challenge set</h3>

challenge_set_df = pd.read_csv('processed_challenge_set.csv')
display(challenge_set_df)

merged_challenge_df = pd.merge(challenge_set_df, summarized_trajectory_df, on='flight_id', how='left')
display(merged_challenge_df)

merged_challenge_df.to_csv("challenge_set_with_trajectories.csv", index=False)
print(f"challenge_set_with_trajectories.csv is saved!")

<h3>Add trajectories to submission set</h3>

submission_set_df = pd.read_csv('processed_submission_set.csv')
display(submission_set_df)

merged_submission_df = pd.merge(submission_set_df, summarized_trajectory_df, on='flight_id', how='left')
display(merged_submission_df)

merged_submission_df.to_csv("submission_set_with_trajectories.csv", index=False)
print(f"submission_set_with_trajectories.csv is saved!")
