==> notebooks/5_label_encode.ipynb <==
from sklearn.preprocessing import LabelEncoder
import pandas as pd
import json

<h3>Append processed submission set to processed challenge set</h3>

challenge_set_df = pd.read_csv('challenge_set_with_trajectories.csv')
display(challenge_set_df)

submission_set_df = pd.read_csv('submission_set_with_trajectories.csv')
display(submission_set_df)

df_combined = pd.concat([challenge_set_df, submission_set_df], ignore_index=True)
display(df_combined)

<h3>label encode them</h3>

def encode(df: pd.DataFrame, categorical_columns: list[str]) -> pd.DataFrame:
    encoded_df = df.copy()
    label_encoder = LabelEncoder()
    for column in encoded_df.columns:
        if column in categorical_columns:     
            encoded_df[column] = label_encoder.fit_transform(df[column])
            if column == 'aircraft_type':
                aircraft_types_encoding = {encoded: original for original, encoded in zip(label_encoder.classes_, range(len(label_encoder.classes_)))}
                with open('aircraft_types_encoding.json', 'w') as json_file:
                    json.dump(aircraft_types_encoding, json_file, indent=4)
    return encoded_df

categorical_columns = [
    'month_day',
    'actual_offblock_time_hour_minute',
    'adep',
    'country_code_adep',
    'arrival_time_hour_minute',
    'ades', 
    'country_code_ades', 
    'aircraft_type', 
    'wtc', 
    'airline', 
]
encoded_df = encode(df=df_combined, categorical_columns=categorical_columns)
display(encoded_df)

<h3>split them back into encoded_challenge_set.csv and encoded_submission_set.csv</h3>

def separate_data(original_df: pd.DataFrame, at_index: int):
    encoded_challenge_set = original_df.iloc[:at_index].reset_index(drop=True)
    encoded_submission_set = original_df.iloc[at_index:].reset_index(drop=True)
    return encoded_challenge_set, encoded_submission_set

encoded_challenge_set, encoded_submission_set = separate_data(original_df=encoded_df, at_index=len(challenge_set_df))

<h3>Save to csv</h3>

display(encoded_challenge_set)

encoded_challenge_set.to_csv("encoded_challenge_set.csv", index=False)
print(f"encoded_challenge_set.csv is saved!")

display(encoded_submission_set)

encoded_submission_set.to_csv("encoded_submission_set.csv", index=False)
print(f"encoded_submission_set.csv is saved!")
